Extract methods for master_key to separate file
