+++ crates/bitwarden/src/crypto/aes_ops.rs
